requirements.txt
streamlit
openai-whisper
pyannote.audio
torch
pydub
nltk
matplotlib
seaborn
ffmpeg-python